fix: Phase 23 specification alignment and test fix (#81)
* fix: align specifications, changelog, and project version for Phase 23, and fix Moran's I test

* fix: address code reviewer findings on version, transactions, and types

+++ src/tabdat/models.py
@@ -624,13 +624,13 @@ class RecodeRange:
   end: float | Literal["max"]
 
 
-RecodeInput = float | int | str | RecodeRange | Literal["missing", "nonmissing", "else"]
+RecodeInput = int | float | str | RecodeRange | Literal["missing", "nonmissing", "else"]
 
 
 @dataclass(frozen=True, config=_MODEL_CONFIG)
 class RecodeRule:
   inputs: tuple[RecodeInput, ...]
-  output: float | int | str
+  output: int | float | str
 
 
 @dataclass(frozen=True, config=_MODEL_CONFIG)
